Workflow for a new tool

+++ tools.mjs
@@ -23,11 +23,16 @@ function toolToReadme(tool) {
     replace(/description: .*/g, `description: ${tool.description}`).
     replace("https://example.com", tool.url);
 }
-function createTool(tool) {
+function createTool(tool, opts={}) {
   const { name, tags } = tool;
   console.log('Creating tool', name);
   const slug = name.toLowerCase().replace(/\s/g, '-');
   const pathname = `gitbook/tools/${slug}`;
+
+  if (!opts.overwrite && fs.existsSync(pathname)) {
+    throw new Error(`Directory ${pathname} already exists`);
+  }
+
   fs.mkdirSync(pathname, { recursive: true });
   fs.writeFileSync(`${pathname}/json.md`, toolToJson(tool));
   fs.writeFileSync(`${pathname}/README.md`, toolToReadme(tool));
@@ -40,7 +45,24 @@ async function createToolOnGitbook(name) {
   console.log("Team and space created. Don't forget to add the team to the space.");
 }
 
+async function findSpace(name) {
+  const response = await fetch('https://api.gitbook.com/v1/collections/jQKvylm6WgaH5IFrlIMh/spaces', {
+    method: 'POST',
+    headers: {
+          "Authorization": `Bearer ${process.env.GITBOOK_API_TOKEN}`
+    },
+  });
+  console.log(response);
+  const data = await response.json();
+  return data.items.find((space) => space.title === name);
+}
+
 async function createSpace(name) {
+  const space = await findSpace(name);
+  if (space) {
+    console.log('Space already exists');
+    return;
+  }
   console.log('Duplicating the template space');
   const response = await fetch('https://api.gitbook.com/v1/spaces/LWUcuebJXer3XFC0YLqM/duplicate', {
     method: 'POST',
@@ -50,10 +72,13 @@ async function createSpace(name) {
   });
   console.log(response);
   const space = await response.json();
+  const updated = await renameSpace(space, name)
   console.log(space);
+}
 
+async function renameSpace(space, name) {
   console.log('Renaming space to', name);
-  await fetch(`https://api.gitbook.com/v1/spaces/${space.id}`, {
+  const response = await fetch(`https://api.gitbook.com/v1/spaces/${space.id}`, {
     method: 'PATCH',
     headers: {
             "Authorization": `Bearer ${process.env.GITBOOK_API_TOKEN}`,
@@ -61,9 +86,29 @@ async function createSpace(name) {
           },
     body: JSON.stringify({ "title": name }),
   });
+  console.log(response);
+  space = await response.json();
+  return space;
+}
+
+async function findTeam(name) {
+  const response = await fetch('https://api.gitbook.com/v1/orgs/WQpOq5ZFue4N6m65QCJq/teams', {
+        method: 'GET',
+        headers: {
+          "Authorization": `Bearer ${process.env.GITBOOK_API_TOKEN}`,
+        }
+  });
+  const data = await response.json();
+  return data.items.find((team) => team.title === name);
 }
 
 async function createTeam(name) {
+  const team = await findTeam(name);
+  if (team) {
+    console.log('Team already exists');
+    return;
+  }
+  console.log('Creating team', name);
   const response = await fetch('https://api.gitbook.com/v1/orgs/WQpOq5ZFue4N6m65QCJq/teams', {
         method: 'PUT',
         headers: {
@@ -73,5 +118,6 @@ async function createTeam(name) {
         body: JSON.stringify({ "title": name }),
   });
   const data = await response.json();
+  console.log(data);
 }
 export default { createTool, createToolOnGitbook };
